+++ src/ResourceType/XBeteiligungPhaseDefinitionCodeMappingResourceConfigBuilder.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 namespace DemosEurope\DemosplanAddon\XBeteiligung\ResourceType;
 
 use DemosEurope\DemosplanAddon\Contracts\Entities\ProcedurePhaseDefinitionInterface;
+use DemosEurope\DemosplanAddon\XBeteiligung\Entity\XBeteiligungDcatApPluStandardCode;
 use DemosEurope\DemosplanAddon\XBeteiligung\Entity\XBeteiligungPhaseDefinitionCodeMapping;
 use EDT\DqlQuerying\Contracts\ClauseFunctionInterface;
 use EDT\DqlQuerying\Contracts\OrderBySortMethodInterface;
@@ -24,6 +25,7 @@ use EDT\JsonApi\ResourceConfig\Builder\MagicResourceConfigBuilder;
  * @template-extends MagicResourceConfigBuilder<ClauseFunctionInterface<bool>, OrderBySortMethodInterface, XBeteiligungPhaseDefinitionCodeMapping>
  *
  * @property-read AttributeConfigBuilderInterface<ClauseFunctionInterface<bool>, XBeteiligungPhaseDefinitionCodeMapping> $xBeteiligungStandardCode
+ * @property-read ToOneRelationshipConfigBuilderInterface<ClauseFunctionInterface<bool>, OrderBySortMethodInterface, XBeteiligungPhaseDefinitionCodeMapping, XBeteiligungDcatApPluStandardCode> $dcatApPluStandardCode
  * @property-read ToOneRelationshipConfigBuilderInterface<ClauseFunctionInterface<bool>, OrderBySortMethodInterface, XBeteiligungPhaseDefinitionCodeMapping, ProcedurePhaseDefinitionInterface> $phaseDefinition
  */
 class XBeteiligungPhaseDefinitionCodeMappingResourceConfigBuilder extends MagicResourceConfigBuilder